+++ libraries/infiniteAverage/CHANGELOG.md
@@ -6,9 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.8] - 2025-05-26
+- improve threshold rounding
+- rename internal variables
+- update examples
+- update readme.md
+- minor edits
+
+
 ## [0.1.7] - 2023-11-06
 - update readme.md
-
 
 ## [0.1.6] - 2022-11-14
 - Add RP2040 support to build-CI.